generate data from range restrictions

+++ src/main/java/pl/poznan/put/sqldatagenerator/generator/OldIntegerAttribute.java
@@ -30,7 +30,7 @@ public class OldIntegerAttribute extends OldAttribute {
     @Override
     protected void generateFromRestriction(boolean negative) {
         IntegerOldRestriction integerRestriction = negative ? (IntegerOldRestriction) getNegativeRestriction() : (IntegerOldRestriction) getRestriction();
-        setValue(RandomGenerator.getInteger(integerRestriction.getRangeSet()));
+        setValue((int) RandomGenerator.getLong(integerRestriction.getRangeSet()));
     }
 
     @Override
